move errors to package errors
